move dashboard project selector to project menu
